refactor(quantization): extract qk256 scalar microcrate
