+++ crates/bitnet-quantization/Cargo.toml
@@ -16,6 +16,7 @@ include = [
 
 [dependencies]
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-qk256-scalar = { path = "../bitnet-qk256-scalar", version = "0.2.1-dev" }
 anyhow.workspace = true
 thiserror.workspace = true
 bytemuck.workspace = true
